+++ js/app-initializer.js
@@ -932,11 +932,17 @@ function initStepInteractiveElements() {
             // Insertar videos según el módulo
             switch (currentModule) {
                 case 'intro':
-                    // Agregar videos en el primer, segundo y último paso del módulo
+                    // Agregar videos en todos los pasos del módulo de introducción
                     if (currentStep === 0) { // Importancia del Análisis de Datos
                         insertIntroVideos(resourcesContainer.id);
                         insertWebsiteLinks(resourcesContainer.id);
                     } else if (currentStep === 1) { // Microsoft Power BI: Componentes Esenciales
+                        insertIntroVideos(resourcesContainer.id);
+                        insertWebsiteLinks(resourcesContainer.id);
+                    } else if (currentStep === 2) { // Recorrido por la Interfaz de Power BI Desktop
+                        insertIntroVideos(resourcesContainer.id);
+                        insertWebsiteLinks(resourcesContainer.id);
+                    } else if (currentStep === 3) { // Tipos de Datos y Modelos (Conceptos Básicos)
                         insertIntroVideos(resourcesContainer.id);
                         insertWebsiteLinks(resourcesContainer.id);
                     } else if (currentStep === 4) { // Asistentes IA en el Flujo de Trabajo